Refactor and redefine test cases for forgot password, product detail, search functionality, and support form
