+++ tests/3-forgotPassword.spec.js
