+++ tests/6-search-functionality.spec.js
